+++ deepinv/physics/mri.py
@@ -1,7 +1,9 @@
 import numpy as np
 import torch
 import torch.fft
-from typing import List, Optional
+from torch import Tensor
+from typing import List, Optional, Union
+import warnings
 from deepinv.physics.forward import DecomposablePhysics
 
 
@@ -22,31 +24,50 @@ class MRI(DecomposablePhysics):
     The complex images :math:`x` and measurements :math:`y` should be of size (B, 2, H, W) where the first channel corresponds to the real part
     and the second channel corresponds to the imaginary part.
 
-    :param torch.Tensor mask: binary mask.
-        The mask size should be of the form (H,W) where H is the image height and W is the image width.
+    A fixed mask can be set at initialisation, or a new mask can be set either at forward (using ``physics(x, mask=mask)``) or using ``update_parameters``.
+
+    :param torch.Tensor mask: binary mask, where 1s represent sampling locations, and 0s otherwise.
+        The mask size can either be (H,W), (C,H,W), or (B,C,H,W) where H, W are the image height and width, C is channels (typically 2) and B is batch size.
     :param torch.device device: cpu or gpu.
 
     |sep|
 
     :Examples:
 
-        Single-coil MRI operator with 4x acceleration:
+        Single-coil accelerated MRI operator with subsampling mask:
 
         >>> from deepinv.physics import MRI
         >>> seed = torch.manual_seed(0) # Random seed for reproducibility
-        >>> x = torch.randn(1, 2, 3, 3) # Define random 3x3 image
-        >>> mask = torch.ones((3, 3))
-        >>> mask[:, ::2] = 0
-        >>> physics = MRI(mask=mask)
+        >>> x = torch.randn(1, 2, 2, 2) # Define random 2x2 image
+        >>> mask = 1 - torch.eye(2) # Define subsampling mask
+        >>> physics = MRI(mask=mask) # Define mask at initialisation
         >>> physics(x)
-        tensor([[-0.5305,  0.0351,  0.3326,  2.1730,  1.7072,  0.0418]])
+        tensor([[[[ 0.0000, -1.4290],
+                [ 0.4564, -0.0000]],
+        <BLANKLINE>
+                [[ 0.0000,  1.8622],
+                [ 0.0603, -0.0000]]]])
+        >>> physics = MRI(img_size=x.shape) # No subsampling
+        >>> physics(x)
+        tensor([[[[ 2.2908, -1.4290],
+                [ 0.4564, -0.1814]],
+        <BLANKLINE>
+                [[ 0.3744,  1.8622],
+                [ 0.0603, -0.6209]]]])
+        >>> physics.update_parameters(mask=mask) # Update mask on the fly
+        >>> physics(x)
+        tensor([[[[ 0.0000, -1.4290],
+                [ 0.4564, -0.0000]],
+        <BLANKLINE>
+                [[ 0.0000,  1.8622],
+                [ 0.0603, -0.0000]]]])
 
     """
 
     def __init__(
         self,
-        mask,
-        img_size=(320, 320),
+        mask: Optional[Tensor] = None,
+        img_size: Optional[tuple] = (320, 320),
         device="cpu",
         **kwargs,
     ):
@@ -54,46 +75,154 @@ class MRI(DecomposablePhysics):
         self.device = device
         self.img_size = img_size
 
-        mask = mask.to(device)
-        if len(mask.shape) == 2:
-            mask = mask.unsqueeze(0).unsqueeze(0)
+        if mask is None:
+            mask = torch.ones(*img_size)
 
-        if mask.shape[1] == 1:
-            mask = torch.cat([mask, mask], dim=1)
+        self.update_parameters(mask=mask.to(self.device))
 
-        self.mask = torch.nn.Parameter(mask, requires_grad=False).to(device)
-
-    def V_adjoint(self, x):  # (B, 2, H, W) -> (B, H, W, 2)
+    def V_adjoint(self, x: Tensor) -> Tensor:  # (B, 2, H, W) -> (B, H, W, 2)
         y = fft2c_new(x.permute(0, 2, 3, 1)).permute(0, 3, 1, 2)
         return y
 
-    def U(self, x):
-        if self.mask.size(0) == 1:
-            return x[:, self.mask[0, ...] > 0]
-        elif x.size(0) == self.mask.size(0):
-            return x[self.mask > 0]
-        else:
+    def U(self, x: Tensor) -> Tensor:
+        if not torch.all((self.mask == 0) | (self.mask == 1)):
+            warnings.warn("The mask must have values 0 or 1")
+
+        if x.size(0) != self.mask.size(0) and self.mask.size(0) != 1:
             raise ValueError(
-                "The batch size of the mask and the input should be the same."
+                "The batch size of the mask and the input should be the same or the mask batch size must be 1."
             )
+        return x * self.mask
 
-    def U_adjoint(self, x):
-        _, c, h, w = self.mask.shape
-        out = torch.zeros((x.shape[0], c, h, w), device=x.device)
+    def U_adjoint(self, x: Tensor) -> Tensor:
+        return self.U(x)
 
-        if self.mask.size(0) == 1:
-            out[:, self.mask[0, ...] > 0] = x
-        elif x.size(0) == self.mask.size(0):
-            out[self.mask > 0] = x
-        else:
-            raise ValueError(
-                "The batch size of the mask and the input should be the same."
-            )
-        return out
-
-    def V(self, x):  # (B, 2, H, W) -> (B, H, W, 2)
+    def V(self, x: Tensor) -> Tensor:  # (B, 2, H, W) -> (B, H, W, 2)
         x = x.permute(0, 2, 3, 1)
         return ifft2c_new(x).permute(0, 3, 1, 2)
+
+    def update_parameters(self, mask=None, check_mask=True, **kwargs):
+        return super().update_parameters(
+            mask=self.check_mask(mask=mask) if check_mask else mask, **kwargs
+        )
+
+    def check_mask(self, mask: Tensor = None) -> None:
+        r"""
+        Updates MRI mask and verifies mask shape to be B,C,H,W.
+
+        :param torch.nn.Parameter, float MRI subsampling mask.
+        """
+        if mask is not None:
+            mask = mask.to(self.device)
+
+            while len(mask.shape) < 4:  # to B,C,H,W
+                mask = mask.unsqueeze(0)
+
+            if mask.shape[1] == 1:  # make complex if real
+                mask = torch.cat([mask, mask], dim=1)
+
+        return mask
+
+
+class DynamicMRI(MRI):
+    r"""
+    Single-coil accelerated dynamic magnetic resonance imaging.
+
+    The linear operator operates in 2D+t videos and is defined as
+
+    .. math::
+
+        y_t = S_t Fx_t
+
+    where :math:`S` applies a time-varying mask, and :math:`F` is the 2D discrete Fourier Transform.
+    This operator has a simple singular value decomposition, so it inherits the structure of
+    :meth:`deepinv.physics.DecomposablePhysics` and thus have a fast pseudo-inverse and prox operators.
+
+    The complex images :math:`x` and measurements :math:`y` should be of size (B, 2, H, W) where the first channel corresponds to the real part
+    and the second channel corresponds to the imaginary part.
+
+    A fixed mask can be set at initialisation, or a new mask can be set either at forward (using ``physics(x, mask=mask)``) or using ``update_parameters``.
+
+    :param torch.Tensor mask: binary mask, where 1s represent sampling locations, and 0s otherwise.
+        The mask size can either be (H,W), (T,H,W), (C,T,H,W) or (B,C,T,H,W) where H, W are the image height and width, T is time-steps, C is channels (typically 2) and B is batch size.
+    :param torch.device device: cpu or gpu.
+
+    |sep|
+
+    :Examples:
+
+        Single-coil accelerated 2D+t MRI operator:
+
+        >>> from deepinv.physics import DynamicMRI
+        >>> seed = torch.manual_seed(0) # Random seed for reproducibility
+        >>> x = torch.randn(1, 2, 2, 2, 2) # Define random video of shape (B,C,T,H,W)
+        >>> mask = torch.rand_like(x) > 0.75 # Define random 4x subsampling mask
+        >>> physics = DynamicMRI(mask=mask) # Physics with given mask
+        >>> physics.update_parameters(mask=mask) # Alternatively set mask on-the-fly
+        >>> physics(x)
+        tensor([[[[[-0.0000,  0.7969],
+                [-0.0000, -0.0000]],
+        <BLANKLINE>
+                [[-0.0000, -1.9860],
+                [-0.0000, -0.4453]]],
+        <BLANKLINE>
+        <BLANKLINE>
+                [[[ 0.0000,  0.0000],
+                [-0.8137, -0.0000]],
+        <BLANKLINE>
+                [[-0.0000, -0.0000],
+                [-0.0000,  1.1135]]]]])
+
+    """
+
+    def flatten(self, a: Tensor) -> Tensor:
+        B, C, T, H, W = a.shape
+        return a.permute(0, 2, 1, 3, 4).reshape(B * T, C, H, W)
+
+    def unflatten(self, a: Tensor, batch_size=1) -> Tensor:
+        BT, C, H, W = a.shape
+        return a.reshape(batch_size, BT // batch_size, C, H, W).permute(0, 2, 1, 3, 4)
+
+    def A(self, x: Tensor, mask: Tensor = None, **kwargs) -> Tensor:
+        mask = self.check_mask(self.mask if mask is None else mask)
+
+        mask_flatten = self.flatten(mask.expand(*x.shape)).to(x.device)
+        y = self.unflatten(
+            super().A(self.flatten(x), mask_flatten, check_mask=False),
+            batch_size=x.shape[0],
+        )
+
+        self.update_parameters(mask=mask, **kwargs)
+        return y
+
+    def A_adjoint(self, y: Tensor, mask: Tensor = None, **kwargs) -> Tensor:
+        mask = self.check_mask(self.mask if mask is None else mask)
+
+        mask_flatten = self.flatten(mask.expand(*y.shape)).to(y.device)
+        x = self.unflatten(
+            super().A_adjoint(self.flatten(y), mask=mask_flatten, check_mask=False),
+            batch_size=y.shape[0],
+        )
+
+        self.update_parameters(mask=mask, **kwargs)
+        return x
+
+    def A_dagger(self, y: Tensor, mask: Tensor = None, **kwargs) -> Tensor:
+        return self.A_adjoint(y, mask=mask, **kwargs)
+
+    def check_mask(
+        self,
+        mask: torch.Tensor = None,
+    ) -> None:
+        r"""
+        Updates MRI mask and verifies mask shape to be B,C,T,H,W.
+
+        :param torch.nn.Parameter, float MRI subsampling mask.
+        """
+        while mask is not None and len(mask.shape) < 5:  # to B,C,T,H,W
+            mask = mask.unsqueeze(0)
+
+        return super().check_mask(mask=mask)
 
 
 #